backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
langchain-groq==0.1.9
langchain-core==0.2.38
langchain==0.2.16
langchain-community==0.2.16
chromadb==0.5.0
pdfplumber==0.11.0
python-multipart==0.0.9
python-dotenv==1.0.0
psycopg2-binary==2.9.9
# ── New dependencies for improved RAG accuracy ──────────────────────────────
# Multilingual sentence embeddings (supports Thai, ~420 MB download on first run)
sentence-transformers>=2.6.0
# BM25 keyword search for hybrid retrieval
rank-bm25>=0.2.2
# HTTP client for evaluate.py
requests>=2.31.0
